+++ src/common/collections/accounts/checks.js
@@ -37,9 +37,9 @@ const _id2index = function( array, id ){
     return -1;
 }
 
-AccountsManager.checks.check_email_address = async function( value, data, opts ){
-    _assert_data_itemrv( 'AccountsManager.checks.check_email_address()', data );
-    //console.debug( 'check_email_address', arguments );
+AccountsManager.checks.email_address = async function( value, data, opts ){
+    _assert_data_itemrv( 'AccountsManager.checks.email_address()', data );
+    //console.debug( 'email_address', arguments );
     let item = data.item.get();
     const index = opts.id ? _id2index( item.emails, opts.id ) : -1;
     if( opts.update !== false ){
@@ -78,8 +78,8 @@ AccountsManager.checks.check_email_address = async function( value, data, opts )
         });
 };
 
-AccountsManager.checks.check_email_verified = async function( value, data, opts ){
-    _assert_data_itemrv( 'AccountsManager.checks.check_email_verified()', data );
+AccountsManager.checks.email_verified = async function( value, data, opts ){
+    _assert_data_itemrv( 'AccountsManager.checks.email_verified()', data );
     const item = data.item.get();
     const index = opts.id ? _id2index( item.emails, opts.id ) : -1;
     if( opts.update !== false ){
@@ -93,14 +93,12 @@ AccountsManager.checks.check_email_verified = async function( value, data, opts 
     return null;
 }
 
-
 // loginAllowed
 //  emit a warning when the user is about to disallow himself
 //  this should nonetheless be prohibited by the UI
-AccountsManager.checks.check_loginAllowed = async function( value, data, opts ){
-    _assert_data_itemrv( 'AccountsManager.checks.check_loginAllowed()', data );
+AccountsManager.checks.loginAllowed = async function( value, data, opts ){
+    _assert_data_itemrv( 'AccountsManager.checks.loginAllowed()', data );
     const item = data.item.get();
-    console.debug( 'value', value );
     if( opts.update !== false ){
         item.loginAllowed = value;
     }
@@ -113,8 +111,8 @@ AccountsManager.checks.check_loginAllowed = async function( value, data, opts ){
     return null;
 };
 
-AccountsManager.checks.check_username = async function( value, data, opts ){
-    _assert_data_itemrv( 'AccountsManager.checks.check_username()', data );
+AccountsManager.checks.username = async function( value, data, opts ){
+    _assert_data_itemrv( 'AccountsManager.checks.username()', data );
     const item = data.item.get();
     if( opts.update !== false ){
         item.username = value;